[CHANGE] Fixed issue with the calibration modal sending extra annotation changed events (#2048)
* Fix issues with the calibration modal sending extra annotation changed events

When we update the scale calibrating we set the Scale value twice to maintain
the precision. However if we call setAnnotationStyles twice this will trigger
two separate annotationChanged events which isn't necessary since we're really
only making one change.

The code that we were using before was also more complicated then it needed to
be because we no longer need to listen to the annotationChanged once after
setting the scale.